+++ tests/testthat/test_normalize_mspct.R
@@ -18,10 +18,17 @@ test_that("source_mspct", {
 
   normalized.mspct <- normalize(my.mspct)
 
-  expect_equal(normalized.mspct[[1]], normalized.mspct[[2]])
-  expect_equal(normalized.mspct[[1]], normalized.mspct[[3]])
-  expect_equal(normalized.mspct[[1]], normalized.mspct[[4]])
-  expect_equal(normalized.mspct[[1]], normalized.mspct[[5]])
+  expect_lt(abs(max(normalized.mspct[[1]]$s.e.irrad) - 1), 1e-10)
+  expect_lt(abs(max(normalized.mspct[[2]]$s.e.irrad) - 1), 1e-10)
+  expect_lt(abs(max(normalized.mspct[[3]]$s.e.irrad) - 1), 1e-10)
+  expect_lt(abs(max(normalized.mspct[[4]]$s.e.irrad) - 1), 1e-10)
+  expect_lt(abs(max(normalized.mspct[[5]]$s.e.irrad) - 1), 1e-10)
+
+  expect_equal(getNormalization(normalized.mspct[[1]])$norm.factor, 1)
+  expect_equal(getNormalization(normalized.mspct[[2]])$norm.factor, 1/2)
+  expect_equal(getNormalization(normalized.mspct[[3]])$norm.factor, 1/3)
+  expect_equal(getNormalization(normalized.mspct[[4]])$norm.factor, 1/4)
+  expect_equal(getNormalization(normalized.mspct[[5]])$norm.factor, 1/5)
 
   expect_equal(class(my.mspct), class(normalized.mspct))