+++ NEWS.md
@@ -4,6 +4,12 @@ editor_options:
     wrap: 72
 ---
 
+# photobiology 0.10.8
+
+-   Update functions `normalize()`, `setNormalized()` and `getNormalized()`,
+    and add new function `getNormalization()`. This changes implement the
+    storage in attribute `normalization` of the operation done.
+    
 # photobiology 0.10.7
 
 -   Add function `ET_ref()` for computation of reference